Replace old distr API by new STDDMaker
The old distr is now replaced by distr2
... and add stdd_left and stdd_right properties to STDDMaker
