+++ stdd_calculator/src/stddc/distr.py
@@ -1,276 +1,341 @@
-"""Calculate the STDP-Streamplot analytically
+#!/usr/bin/env python3
 
-Calc the STDP-Streamplot of a two neuron buesing network with arbitrary PSP
-shape analytically.
-
-Detailed description tba
-"""
-
-import logging
-import sys
+import warnings
+from typing import Callable, Tuple, Union
 
 import matplotlib.pyplot as plt
 import numpy as np
-from scipy import sparse as sm
+import numpy.typing as npt
+from scipy import sparse
 from scipy.sparse.linalg import eigs
 
+from .funcs import rect_PSP
 from .matrix import transition_matrix
 
-logging.basicConfig(
-    format="%(asctime)s %(levelname)s: %(message)s",
-    level=logging.INFO,
-    datefmt="%d/%m/%Y %H:%M:%S",
-)
+PSPFunc = Callable[[int, int, int], float]
+
+SalFunc = Callable[[npt.NDArray, Union[int, float], ...], npt.NDArray]
 
 
-def rect_PSP(zeta, tau_ref, tau_syn):
-    """rectangular psp shape"""
-    return np.heaviside(-zeta + tau_ref, 0.0) * np.heaviside(zeta, 1.0)
+class EigenvalueError(Exception):
+    """Custom exception for eigenvalue-related errors."""
+
+    pass
 
 
-def exp_PSP(zeta, tau_ref, tau_syn):
-    """exponential psp shape"""
-    a = (tau_ref / tau_syn) / (1 - np.exp(-tau_ref / tau_syn))
-    return a * np.exp(-zeta / tau_syn)
+class STDDMaker:
+    """Docstring."""
 
-
-def cutoff_PSP(zeta, tau_ref, tau_syn):
-    if zeta <= tau_ref:
-        a = (tau_ref / tau_syn) / (1 - np.exp(-tau_ref / tau_syn))
-        return a * np.exp(-zeta / tau_syn)
-    else:
-        return 0.0
-
-
-def tail_PSP(zeta, tau_ref, tau_syn):
-    if zeta <= tau_ref:
-        return 1.0
-    else:
-        a = (tau_ref / tau_syn) / (1 - np.exp(-tau_ref / tau_syn))
-        return a * np.exp(-zeta / tau_syn)
-
-
-def alpha_PSP(zeta, tau_ref, tau_syn):
-    if zeta > 0.0:
-        return tau_ref / tau_syn**2 * zeta * np.exp(-zeta / tau_syn)
-    else:
-        return 0.0
-
-
-def create_transition_matrix(PSP, tau_ref, t_max, w_12, w_21, b_1, b_2, tau_syn):
-    """Create a sparse transition matrix"""
-    T_sparse = sm.dok_matrix((t_max**2, t_max**2))
-
-    for i, j, res in transition_matrix(
-        PSP, tau_ref, t_max, w_12, w_21, b_1, b_2, tau_syn
+    def __init__(
+        self,
+        psp: PSPFunc,
+        t_max: int,
+        t_ref: int,
+        t_syn: int,
+        w_12: float,
+        w_21: float,
+        b_1: float,
+        b_2: float,
     ):
-        T_sparse[i, j] = res
-    logging.info(
-        "Size of sparse matrix in MB {}".format(sys.getsizeof(T_sparse) / 1.0e6)
-    )
+        """
+        Initialize the STDDMaker with the given parameters.
 
-    return T_sparse
+        :param psp: A function representing the Post-Synaptic Potential.
+        :type psp: PSPFunc
+        :param t_max: Maximum time step.
+        :type t_max: int
+        :param t_ref: Refractory period.
+        :type t_ref: int
+        :param t_syn: Synaptic time constant.
+        :type t_syn: int
+        :param w_12: Synaptic weight from neuron 1 to neuron 2.
+        :type w_12: float
+        :param w_21: Synaptic weight from neuron 2 to neuron 1.
+        :type w_21: float
+        :param b_1: Bias for neuron 1.
+        :type b_1: float
+        :param b_2: Bias for neuron 2.
+        :type b_2: float
 
+        The PSPFunc is expected to be a callable that takes three integers and
+        returns a float.
+        """
+        assert t_ref < t_max, "t_ref must be smaller than t_max."
+        self.psp = psp
+        self.t_max = t_max
+        self.t_ref = t_ref
+        self.t_syn = t_syn
+        self.w_21 = w_12
+        self.w_12 = w_21
+        self.b_2 = b_1
+        self.b_1 = b_2
 
-def calc_equilibrium_distr(T_sparse):
-    """Calculate the equilibrium distribution for the given transition matrix
+        # construct two matrices, one that calcs the right half...
+        self.T_full_right = None
+        # and by exchanging the neurons the left half...
+        self.T_full_left = None
 
-    Calculates the first eigen value near 1.+0i and the corresponding eigen
-    vector which is equal to the equilibrium distribution of the two neuron
-    system.
-    """
-    eig_val, eig_vec = eigs(T_sparse, k=1, sigma=1.0 + 0.0j)
-    logging.info(("eigen value:", eig_val))
-    # find eigen value = 1.
-    equil_id = np.where(np.isclose(eig_val, 1.0 + 0.0j))[0][0]
-    equil_distr = np.real(eig_vec[:, equil_id]).flatten()
-    all_real = np.all(np.imag(equil_distr) == 0.0)
-    logging.info(("All values of the eigen vector are real:", all_real))
-    equil_distr /= np.sum(equil_distr)
+        self.stdd = None
 
-    return equil_distr
+        self.times = np.arange(-t_max + 1, t_max)
 
+    @property
+    def times_left(self) -> np.ndarray:
+        """Time axis for the left half of the STDD (dt < 0)."""
+        return np.arange(-self.t_max + 1, 0)
 
-def create_nonspiking_trans_matrix(T_sparse, t_max):
-    """
-    Returns a copy of the transition matrix where all entries are zero that
-    would let either neuron spike.
-    """
-    T_tilde = T_sparse.copy()
+    @property
+    def times_right(self) -> np.ndarray:
+        """Time axis for the right half of the STDD (dt > 0)."""
+        return np.arange(1, self.t_max)
 
-    # neuron 2 cannot spike
-    ceta_2 = 0
-    for ceta_1 in range(1, t_max):
-        i = ceta_1 * t_max + ceta_2
-        T_tilde[i, :] = 0.0
+    @property
+    def stdd_left(self) -> np.ndarray:
+        """Left half of the STDD (dt < 0), aligned with times_left."""
+        if self.stdd is None:
+            raise RuntimeError("Call calc_stdd() first.")
+        return self.stdd[: self.t_max - 1]
 
-    # neuron 1 cannot spike
-    ceta_1 = 0
-    for ceta_2 in range(1, t_max):
-        i = ceta_1 * t_max + ceta_2
-        T_tilde[i, :] = 0.0
+    @property
+    def stdd_right(self) -> np.ndarray:
+        """Right half of the STDD (dt > 0), aligned with times_right."""
+        if self.stdd is None:
+            raise RuntimeError("Call calc_stdd() first.")
+        return self.stdd[self.t_max :]
 
-    # print(T_tilde.toarray())
-    T_tilde = T_tilde.tocsr()
-    return T_tilde
+    def create_transition_matrix(
+        self, w_12: float, w_21: float, b_1: float, b_2: float
+    ) -> sparse.dok_matrix:
+        """
+        Create a transition matrix.
 
+        :param psp: Function representing the Post-Synaptic Potential.
+        :type psp: PSPFunc
+        :param t_ref: Refractory period.
+        :type t_ref: int
+        :param t_max: Maximum time step.
+        :type t_max: int
+        :param w_12: Synaptic weight from neuron 1 to neuron 2.
+        :type w_12: float
+        :param w_21: Synaptic weight from neuron 2 to neuron 1.
+        :type w_21: float
+        :param b_1: Bias for neuron 1.
+        :type b_1: float
+        :param b_2: Bias for neuron 2.
+        :type b_2: float
+        :param t_syn: Synaptic time constant.
+        :type t_syn: int
+        :return: A sparse transition matrix.
+        :rtype: sparse.dok_matrix
+        """
+        T_full = sparse.dok_matrix((self.t_max**2, self.t_max**2))
 
-def create_trans_matrix_neuron_2(T_sparse, t_max):
-    """Returns a version of T where neuron 2 will spike.
+        # Assuming transition_matrix is a function that yields tuples of (i, j, val)
+        for i, j, val in transition_matrix(
+            self.psp, self.t_ref, self.t_max, w_12, w_21, b_1, b_2, self.t_syn
+        ):
+            T_full[i, j] = val
 
-    Used to calc the RIGHT side of the dt-distribution (dt > 0)
-    The dimention of the matrix is (t_max, t_max**2) --> rho = T * q, with rho
-    being the desired dt-distribution.
-    """
-    T_hat_right = sm.dok_matrix((t_max, t_max**2))
-    for ceta_1 in range(t_max):
-        # if neuron 2 spikes, copy line into T_hat
-        i = ceta_1 * t_max
-        T_hat_right[ceta_1, :] = T_sparse[i, :]
-    # print(T_hat.toarray())
-    T_hat_right = T_hat_right.tocsr()
-    return T_hat_right
+        return T_full
 
+    @staticmethod
+    def calc_equilibrium_distr(
+        T_full: Union[sparse.csr_matrix, sparse.csc_matrix],
+    ) -> np.ndarray:
+        """
+        Calculate the equilibrium distribution for the given transition matrix.
 
-def create_trans_matrix_neuron_1(T_sparse, t_max):
-    """Returns a version of T where neuron 1 will spike.
+        Calculates the first eigenvalue near 1.+0i and the corresponding
+        eigenvector which is equal to the equilibrium distribution of the
+        two neuron system.
 
-    Used to calc the LEFT side of the dt-distribution (dt < 0)
-    The dimention of the matrix is (t_max, t_max**2) --> rho = T * q, with rho
-    being the desired dt-distribution.
-    """
-    T_hat_left = sm.dok_matrix((t_max, t_max**2))
-    for ceta_1 in range(t_max):
-        # if neuron 1 spikes, copy line into T_hat
-        T_hat_left[ceta_1, :] = T_sparse[ceta_1, :]
-    # print(T_hat.toarray())
-    T_hat_left = T_hat_left.tocsr()
-    return T_hat_left
+        :param T_full: Sparse transition matrix.
+        :type T_full: Union[scipy.sparse.csr_matrix, scipy.sparse.csc_matrix]
+        :return: Equilibrium distribution vector.
+        :rtype: numpy.ndarray
+        """
+        try:
+            eig_val, eig_vec = eigs(T_full, k=1, sigma=1.0 + 0.0j)
+        except RuntimeError:
+            warnings.warn(
+                "T_full seems to be singular (or numerically close to "
+                "singularity). Add regularization and retry.",
+                RuntimeWarning,
+            )
+            reg = sparse.eye(T_full.shape[0]) * 1e-10
+            eig_val, eig_vec = eigs(T_full + reg, k=1, sigma=1.0 + 0.0j)
 
+        # Find eigenvalue = 1.
+        equil_id = np.where(np.isclose(eig_val, 1.0 + 0.0j))[0][0]
+        if not equil_id.size == 1:
+            raise EigenvalueError(
+                f"Expected one eigenvalue with 1.0 + 0.0j, but found {eig_val}."
+            )
 
-def calc_half_histogram(distr, T_tilde, T_hat, t_max):
-    """Calculate one half of the dt distribution
+        equil_distr = np.real(eig_vec[:, equil_id]).flatten()
 
-    Takes a starting distribution, which determines if the left or right side
-    of the histogram will be calculated.
-    """
-    dt_hist = np.zeros(t_max)
+        if not np.all(np.imag(equil_distr) == 0.0):
+            raise EigenvalueError(
+                f"Expected eigenvector to be real, but found {equil_distr}"
+            )
 
-    # calc the distribution of next time step when we allow neuron 2 or 1 to
-    # spike:
-    distr_spike = T_hat.dot(distr)
-    dt_hist += distr_spike
+        equil_distr /= np.sum(equil_distr)
 
-    for t in range(t_max):
-        # new distribution after one update where spikes are forbidden
-        distr = T_tilde.dot(distr)
-        # calc the distribution of next time step when we allow neuron 2 to
-        # spike:
-        distr_spike = T_hat.dot(distr)
-        dt_hist += distr_spike
+        return equil_distr
 
-    return dt_hist
+    @staticmethod
+    def _create_nonspiking_trans_matrix(
+        T_full: Union[sparse.csr_matrix, sparse.csc_matrix], t_max: int
+    ) -> sparse.csr_matrix:
+        """
+        Returns a copy of the transition matrix where all entries are zero that
+        would let either neuron spike.
+        """
+        T_tilde = T_full.copy()
 
+        # neuron 2 cannot spike
+        ceta_2 = 0
+        for ceta_1 in range(1, t_max):
+            i = ceta_1 * t_max + ceta_2
+            T_tilde[i, :] = 0.0
 
-def calc_dt_distr(PSP, tau_ref, t_max, w_12, w_21, b_1, b_2, tau_syn):
-    """Calculates the complete histogram of the 2 neuron system
+        # neuron 1 cannot spike
+        ceta_1 = 0
+        for ceta_2 in range(1, t_max):
+            i = ceta_1 * t_max + ceta_2
+            T_tilde[i, :] = 0.0
 
-    Takes the parameters of the network (tau_ref, t_max, w_12, w_21, b_1, b_2, tau_syn)
-    and the PSP shape (function) and calculates the whole distribution of the
-    interpike intervalls between two neurons.
-    """
-    # Create transition matrix:
-    logging.info("Create transition matrix...")
-    T = create_transition_matrix(PSP, tau_ref, t_max, w_12, w_21, b_1, b_2, tau_syn)
+        T_tilde = T_tilde.tocsr()
+        return T_tilde
 
-    # Create non-spiking transition matrix:
-    logging.info("Create non spiking transition matrix.")
-    T_nospike = create_nonspiking_trans_matrix(T, t_max)
+    def _create_trans_matrix_neuron_2(
+        self, T_full: Union[sparse.csr_matrix, sparse.csc_matrix]
+    ) -> sparse.csr_matrix:
+        """Docstring."""
+        T_hat_right = sparse.dok_matrix((self.t_max, self.t_max**2))
+        for ceta_1 in range(1, self.t_max):
+            # if neuron 2 spikes, copy line into T_hat
+            i = ceta_1 * self.t_max
+            j = (ceta_1 - 1) * self.t_max + self.t_ref - 1
+            k = (ceta_1 - 1) * self.t_max + self.t_max
+            T_hat_right[ceta_1, j:k] = T_full[i, j:k] + T_full[0, j:k]
 
-    # Create transition matrix for right side:
-    logging.info("Create transition matrix for right side.")
-    T_right = create_trans_matrix_neuron_2(T, t_max)
+        T_hat_right = T_hat_right.tocsr()
+        return T_hat_right
 
-    # Create transition matrix for right side:
-    logging.info("Create transition matrix for left side.")
-    T_left = create_trans_matrix_neuron_1(T, t_max)
+    def _calc_half_histogram(
+        self, T_full: Union[sparse.csr_matrix, sparse.csc_matrix]
+    ) -> np.ndarray:
+        """Returns a version of T where neuron 2 will spike.
 
-    # Calculate equilibrium distribution:
-    logging.info("Calc the equilibrium distribution.")
-    equil_distr = calc_equilibrium_distr(T)
+        Used to calc the RIGHT side of the dt-distribution (dt > 0)
+        The dimention of the matrix is (t_max, t_max**2) --> rho = T * q, with rho
+        being the desired dt-distribution.
+        """
+        # create the transition_matrix when no spike occurs
+        T_nospike = self._create_nonspiking_trans_matrix(T_full, self.t_max)
+        # create the transition_matrix when the second neuron spikes
+        T_spike = self._create_trans_matrix_neuron_2(T_full)
 
-    # Calc the right side of the histogram:
-    logging.info("Calc right side of histogram")
-    distr_right = np.copy(equil_distr)
-    # set initial distribution to states in which neuron one spikes:
-    distr_right[t_max:] = 0.0
-    # renormalize distribution:
-    distr_right /= np.sum(distr_right)
+        # calculate the equilibrium distribution
+        equil_distr = self.calc_equilibrium_distr(T_full)
+        # set initial distribution to states in which neuron one spikes:
+        equil_distr[self.t_max :] = 0.0
+        equil_distr /= np.sum(equil_distr)
 
-    hist_right = calc_half_histogram(distr_right, T_nospike, T_right, t_max)
+        stdd = np.zeros(self.t_max)
 
-    # Calc the left side of the histogram:
-    logging.info("Calc left side of histogram")
-    distr_left = np.zeros_like(equil_distr)
-    # set initial distribution to states in which neuron two spikes:
-    for i in range(t_max):
-        distr_left[i * t_max] = equil_distr[i * t_max]
-    # renormalize distribution:
-    distr_left /= np.sum(distr_left)
+        # Calculate the distribution of the next time step when we allow neuron 2
+        # or 1 to spike
+        distr_spike = T_spike.dot(equil_distr)
+        stdd += distr_spike
 
-    hist_left = calc_half_histogram(distr_left, T_nospike, T_left, t_max)
+        for t in range(self.t_max):
+            # New distribution after one update where spikes are forbidden
+            equil_distr = T_nospike.dot(equil_distr)
+            # Calculate the distribution of the next time step when we allow
+            # neuron 2 to spike
+            distr_spike = T_spike.dot(equil_distr)
+            stdd += distr_spike
 
-    dt_left_norm = hist_left / np.sum(hist_left) * 0.5
-    dt_right_norm = hist_right / np.sum(hist_right) * 0.5
-    dt_hist_norm = np.concatenate((dt_left_norm[:0:-1], dt_right_norm))
+        # cut off the last bit, because it contains all the remaining transitions
+        # (i.e. an artefact comming from the existanance of a t_max)
+        # stdd = stdd[:-1]
+        # normalize stdd:
+        stdd /= np.sum(stdd) * 2.0
 
-    return dt_hist_norm
+        return stdd
+
+    def calc_right(self):
+        """DOCSTRING."""
+
+        self.T_full_right = self.create_transition_matrix(
+            self.w_12, self.w_21, self.b_1, self.b_2
+        )
+
+        stdd_right = self._calc_half_histogram(self.T_full_right)
+        return stdd_right[1:]
+
+    def calc_left(self):
+        """DOCSTRING."""
+
+        self.T_full_left = self.create_transition_matrix(
+            self.w_21, self.w_12, self.b_2, self.b_1
+        )
+
+        stdd_left = self._calc_half_histogram(self.T_full_left)
+        return stdd_left[:0:-1]
+
+    def calc_stdd(self, fill_middle: str = "nan") -> npt.NDArray:
+        """DOCSTRING."""
+        ALLOWED = ["nan", "zero", "smooth"]
+        assert (
+            fill_middle in ALLOWED
+        ), f"optional argument 'fill_middle' must be from {ALLOWED}"
+
+        stdd_right = self.calc_right()
+        stdd_left = self.calc_left()
+
+        if fill_middle == "nan":
+            fill_value = np.array([np.nan])
+        elif fill_middle == "zero":
+            fill_value = np.array([0.0])
+        else:
+            a = (stdd_left[-1] + stdd_right[0]) * 0.5
+            fill_value = np.array([a])
+
+        self.stdd = np.concatenate([stdd_left, fill_value, stdd_right])
+
+        return np.copy(self.stdd)
+
+    def calc_sal(self, func: SalFunc, *args: Union[int, float]) -> Tuple[float, float]:
+        """DOCSTRING."""
+        kernel_vals = func(self.times, *args)
+
+        if self.stdd is None:
+            stdd = self.calc_stdd(fill_middle="zero")
+        else:
+            stdd = np.copy(self.stdd)
+            stdd[self.t_max - 1] = 0.0
+
+        sal_12 = np.sum(stdd * kernel_vals)
+        sal_21 = np.sum(stdd * kernel_vals[::-1])
+
+        return sal_12, sal_21
 
 
 def main():
     """DOCSTRING."""
-    PSP = exp_PSP
-    logging.info("No simulation data.")
-    tau_bio = 10.0
-    tau_ref = 50
-    tau_syn = 25
-    t_max = 150
-    w_12 = 0.3
-    w_21 = -0.7
-    b_1 = 1.3
-    b_2 = 0.2
-    logging.info("")
-    logging.info(
-        "\n".join(
-            [
-                "tau_ref = {}, t_max = {}".format(tau_ref, t_max),
-                "b_1 = {}, b_2 = {}".format(b_1, b_2),
-                "w_12 = {}, w_21 = {}".format(w_12, w_21),
-            ]
-        )
+    stdd_maker = STDDMaker(
+        psp=rect_PSP, t_max=30, t_ref=10, t_syn=10, w_12=0.5, w_21=1.0, b_1=0.0, b_2=0.0
     )
-
-    hist_analytical = calc_dt_distr(PSP, tau_ref, t_max, w_12, w_21, b_1, b_2, tau_syn)
-
-    hist_analytical *= tau_ref / tau_bio
-    logging.info(str(hist_analytical))
-    xs = np.arange(-t_max + 1, t_max) / tau_ref * tau_bio
+    _ = stdd_maker.calc_stdd()
 
     fig, ax = plt.subplots()
-    ax.plot(xs, hist_analytical, label="matrix method")
-    ax.set_xlim(-30, 30)
-    ax.set_xlabel(r"$\Delta t_{pp}$ [ms]")
-    ax.set_ylabel(r"$\rho(\Delta t$")
-    ax.legend()
-    ax.grid()
-    ax.set_title(
-        "b1={}, b_2={}, w12={}, w21={}, tau_ref={}".format(
-            b_1, b_2, w_12, w_21, tau_ref
-        )
-    )
-    file_name = "histogram.png"
-    fig.savefig(file_name, dpi=150)
-    logging.info("Histogram saved to {}".format(file_name))
+    ax.plot(stdd_maker.times, stdd_maker.stdd)
+    ax.set_xlabel("Delta t")
+    ax.set_ylabel("STDD")
+    fig.savefig("stdd.png")
 
 
 if __name__ == "__main__":
